Privilèges associés à la gestion des qualités des membres du jury

+++ module/Soutenance/src/Soutenance/Controller/QualiteController.php
@@ -2,8 +2,10 @@
 
 namespace Soutenance\Controller;
 
+use BjyAuthorize\Exception\UnAuthorizedException;
 use Soutenance\Entity\Qualite;
 use Soutenance\Form\QualiteEdition\QualiteEditionForm;
+use Soutenance\Provider\Privilege\QualitePrivileges;
 use Soutenance\Service\Membre\MembreServiceAwareTrait;
 use Zend\Http\Request;
 use Zend\Mvc\Controller\AbstractActionController;
@@ -17,6 +19,12 @@ class QualiteController extends AbstractActionController
      * Affiche la liste des qualités enregistrées dans SyGAL et permet l'accés aux fonctions d'ajout, d'édition et de retrait
      */
     public function indexAction() {
+
+        $isAllowed = $this->isAllowed(QualitePrivileges::getResourceId(QualitePrivileges::SOUTENANCE_QUALITE_VISUALISER));
+        if (!$isAllowed) {
+            throw new UnAuthorizedException("Vous êtes non authorisé(e) à visualiser la liste des qualités affectables aux membres du jury.");
+        }
+
         $qualites = $this->getMembreService()->findAllQualites();
 
         return new ViewModel([
@@ -25,6 +33,12 @@ class QualiteController extends AbstractActionController
     }
 
     public function editerAction() {
+
+        $isAllowed = $this->isAllowed(QualitePrivileges::getResourceId(QualitePrivileges::SOUTENANCE_QUALITE_MODIFIER));
+        if (!$isAllowed) {
+            throw new UnAuthorizedException("Vous êtes non authorisé(e) à modifier la liste des qualités affectables aux membres du jury.");
+        }
+
         /** @var Qualite $qualite */
         $idQualite = $this->params()->fromRoute('qualite');
         $qualite = null;
@@ -36,7 +50,7 @@ class QualiteController extends AbstractActionController
 
         /** @var QualiteEditionForm $form */
         $form = $this->getServiceLocator()->get('FormElementManager')->get(QualiteEditionForm::class);
-        $form->setAttribute('action', $this->url()->fromRoute('soutenance/qualite/editer', ['qualite' => $idQualite], [], true));
+        $form->setAttribute('action', $this->url()->fromRoute('qualite/editer', ['qualite' => $idQualite], [], true));
         $form->bind($qualite);
 
         /** @var Request $request */
@@ -59,13 +73,19 @@ class QualiteController extends AbstractActionController
     }
 
     public function effacerAction() {
+
+        $isAllowed = $this->isAllowed(QualitePrivileges::getResourceId(QualitePrivileges::SOUTENANCE_QUALITE_MODIFIER));
+        if (!$isAllowed) {
+            throw new UnAuthorizedException("Vous êtes non authorisé(e) à modifier la liste des qualités affectables aux membres du jury.");
+        }
+
         /** @var Qualite $qualite */
         $idQualite = $this->params()->fromRoute('qualite');
         $qualite = $this->getMembreService()->getQualiteById($idQualite);
 
         $this->getMembreService()->removeQualite($qualite);
 
-        $this->redirect()->toRoute('soutenance/qualite', [], [], true);
+        $this->redirect()->toRoute('qualite', [], [], true);
     }
 
 }
